refacto: merge page and page-loader
