+++ portal-front/app/(application)/(admin)/admin/parameters/page.tsx
@@ -1,18 +1,34 @@
 'use client';
 import GuardCapacityComponent from '@/components/admin-guard';
+import { Parameters } from '@/components/admin/parameters/parameters';
+import { BreadcrumbNav } from '@/components/ui/breadcrumb-nav';
+import { UseTranslationsProps } from '@/i18n/config';
 import { RESTRICTION } from '@/utils/constant';
+import { useTranslations } from 'next-intl';
 import * as React from 'react';
-import PageLoader from './page-loader';
 
 interface PageProps {}
 
 // Component
 const Page: React.FunctionComponent<PageProps> = () => {
+  const t = useTranslations();
+
+  const breadcrumbValue = (t: UseTranslationsProps) => [
+    {
+      label: t('MenuLinks.Settings'),
+    },
+    {
+      label: t('MenuLinks.Parameters'),
+    },
+  ];
+
   return (
     <GuardCapacityComponent
       displayError
       capacityRestriction={[RESTRICTION.CAPABILITY_BYPASS]}>
-      <PageLoader />
+      <BreadcrumbNav value={breadcrumbValue(t)} />
+      <h1 className="pb-s">{t('MenuLinks.Parameters')}</h1>
+      <Parameters />
     </GuardCapacityComponent>
   );
 };
